add particle custom wrapper
