+++ client/package.json
@@ -8,15 +8,15 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "bootstrap": "^5.1.3",
     "graphql": "^15.4.0",
     "jwt-decode": "^3.1.2",
     "react": "^17.0.2",
+    "react-bootstrap": "^2.0.3",
     "react-dom": "^17.0.1",
-    "react-particles-js": "^3.6.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
-    "react-tsparticles": "^1.37.5",
-    "tsparticles": "^1.37.5"
+    "react-tsparticles": "^1.37.5"
   },
   "scripts": {
     "start": "react-scripts start",
