+++ client/src/utils/ParticleWrapper/particle.css
@@ -1,0 +1,20 @@
+.particleParent{
+    position: relative;
+    z-index: 0;
+    margin: 0 0 0.5rem;
+}
+
+.particleChild{
+    position: absolute;
+    width: 100%;
+    z-index: 1;
+    height: 100%;
+}
+
+#tsparticles{
+    position: absolute;
+    width: 100%;
+    z-index: 0;
+    margin: 0 0 0.5rem !important;
+}
+
