--- client/src/utils/particle.css
@@ -1,16 +1,0 @@
-.particleParent{
-    position: relative;
-    z-index: 0;
-}
-
-.particleChild{
-    z-index: 1;
-    position: absolute;
-    top: 10vh;
-    color: white
-}
-
-.tsparticles-canvas-el{
-    z-index: 0;
-}
-
